techpack: audio: Construct and add audio techpack Kconfig to build
This is the only way to solve the mess that happens whenever
we try to rebuild the kernel without mrproper'ing it first.

It's the most dirty thing on Earth, but it's really the only
possible option.

Signed-off-by: xiaoleGun <1592501605@qq.com>
(cherry picked from commit 04398bec5d7a21c6b4c7103f2cb960cbedd19e1e)