techpack: audio: q6voice: Force AVCS CVP version 1 on legacy SoC
Force the AVCS version to CVP_VERSION_1 for all of the AVCS
services on sdm636: It cannot execute the Q6 Service Version
command on it's Q6 DSP because of older firmware not supporting it.
We know that for this SoC the version is always 1, hence
forcing it is totally safe.

Signed-off-by: xiaoleGun <1592501605@qq.com>
(cherry picked from commit af4da29842b7a679055bf5d74e3bae9056588db3)